requirements.txt
# FastAPI and server
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
jinja2>=3.1.3

# LINE SDK
line-bot-sdk>=3.5.0

# LLM Providers
openai>=1.12.0
google-generativeai>=0.4.0
ollama>=0.1.6

# RAG and Vector Store
langchain>=0.1.6
langchain-core>=0.1.0
langchain-text-splitters>=0.0.1
langchain-community>=0.0.19
langchain-openai>=0.0.5
langchain-huggingface>=0.0.1
chromadb>=0.4.22
sentence-transformers>=2.3.1

# Document Processing
pypdf>=4.0.1
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.1
pydantic>=2.6.0
pydantic-settings>=2.1.0
aiohttp>=3.9.3
httpx>=0.26.0

# Database (optional, for conversation history)
sqlalchemy>=2.0.25
aiosqlite>=0.19.0
